# api/openapi.policy.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
    "/api/v1/admin/collection-exceptions" : {
      "get" : {
        "description" : "최근 처리 시각 순. 재처리에 성공한 항목은 제외하며, 같은 시각에는 수집 요청 순번 역순·항목 위치 순으로 정렬한다.",
        "operationId" : "listCollectionExceptions",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1,
            "maximum" : 1000,
            "minimum" : 1
          }
        }, {
          "in" : "query",
          "name" : "pageSize",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20,
            "maximum" : 50,
            "minimum" : 1
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminSliceCollectionExceptionItem"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "조회 조건 오류"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "로그인 필요"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "권한 또는 CSRF 토큰 확인 필요"
          },
          "503" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "수집 처리 또는 이력 조회 실패"
          }
        },
        "security" : [ {
          "memberSession" : [ ]
        } ],
        "summary" : "관리자 전용 수집 항목 검증·저장 실패 목록",
        "tags" : [ "collection-exception-controller" ]
      }
    },
    "/api/v1/admin/collection-exceptions/pages" : {
      "get" : {
        "description" : "수집 요청 순번 역순. 정상 처리된 페이지는 제외한다. 원본 응답과 임의 오류 문자열을 노출하지 않으며 재처리를 실행하지 않는다.",
        "operationId" : "listCollectionPageFailures",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1,
            "maximum" : 1000,
            "minimum" : 1
          }
        }, {
          "in" : "query",
          "name" : "pageSize",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20,
            "maximum" : 50,
            "minimum" : 1
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminSliceCollectionPageFailure"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "조회 조건 오류"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "로그인 필요"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "권한 또는 CSRF 토큰 확인 필요"
          },
          "503" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "수집 처리 또는 이력 조회 실패"
          }
        },
        "security" : [ {
          "memberSession" : [ ]
        } ],
        "summary" : "관리자 전용 페이지 수집 실패 목록",
        "tags" : [ "collection-exception-controller" ]
      }
    },
    "/api/v1/admin/collection-exceptions/replays" : {
      "get" : {
        "operationId" : "listCollectionReplays",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1,
            "maximum" : 1000,
            "minimum" : 1
          }
        }, {
          "in" : "query",
          "name" : "pageSize",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20,
            "maximum" : 50,
            "minimum" : 1
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminSliceCollectionReplayResult"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "조회 조건 오류"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "로그인 필요"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "권한 또는 CSRF 토큰 확인 필요"
          },
          "503" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "수집 처리 또는 이력 조회 실패"
          }
        },
        "security" : [ {
          "memberSession" : [ ]
        } ],
        "summary" : "관리자 수집 항목 재처리 이력",
        "tags" : [ "collection-exception-controller" ]
      }
    },
    "/api/v1/admin/collection-exceptions/{runId}/{itemIndex}" : {
      "get" : {
        "description" : "저장된 실패 유형만 제공하며 구체적인 실패 원인을 추정하지 않는다. 원본 수정이나 재처리를 실행하지 않는다.",
        "operationId" : "getCollectionException",
        "parameters" : [ {
          "in" : "path",
          "name" : "runId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "path",
          "name" : "itemIndex",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 9,
            "minimum" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CollectionExceptionDetail"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "조회 조건 오류"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "로그인 필요"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "권한 또는 CSRF 토큰 확인 필요"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "현재 실패 상태인 항목 없음"
          },
          "503" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "수집 처리 또는 이력 조회 실패"
          }
        },
        "security" : [ {
          "memberSession" : [ ]
        } ],
        "summary" : "관리자 전용 수집 실패 원본·현재 정책 조회",
        "tags" : [ "collection-exception-controller" ]
      }
    },
    "/api/v1/admin/collection-exceptions/{runId}/{itemIndex}/replays" : {
      "post" : {
        "description" : "같은 요청 ID·항목·작업자·사유·처리 횟수는 기존 결과를 반환한다. 원본 수정이나 외부 수집을 실행하지 않는다.",
        "operationId" : "replayCollectionItem",
        "parameters" : [ {
          "in" : "path",
          "name" : "runId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "path",
          "name" : "itemIndex",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 9,
            "minimum" : 0
          }
        }, {
          "description" : "GET /api/v1/session에서 받은 csrfToken",
          "in" : "header",
          "name" : "X-CSRF-TOKEN",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CollectionReplayRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CollectionReplayResult"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "조회 조건 오류"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "로그인 필요"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "권한 또는 CSRF 토큰 확인 필요"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "처리 상태 변경 또는 저장 원본 재처리 불가"
          },
          "503" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "수집 처리 또는 이력 조회 실패"
          }
        },
        "security" : [ {
          "memberSession" : [ ]
        } ],
        "summary" : "관리자 사유를 남기고 저장 원본의 실패 항목 하나를 재처리",
        "tags" : [ "collection-exception-controller" ]
      }
    },
...
    "/api/v1/admin/policy-ai-runs" : {
      "get" : {
        "description" : "요청별 마지막 자동 시도를 최근 순으로 조회한다. 검색·필터 후 페이지를 나누며 조회로 실행·재호출·정산하지 않는다.",
        "operationId" : "listPolicyAiRuns",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1,
            "maximum" : 1000,
            "minimum" : 1
          }
        }, {
          "in" : "query",
          "name" : "pageSize",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20,
            "maximum" : 50,
            "minimum" : 1
          }
        }, {
          "in" : "query",
          "name" : "filter",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "ALL",
            "enum" : [ "ALL", "RUNNING", "COMPLETED", "RETRY_PENDING", "INTERRUPTED", "REVIEW_REQUIRED", "SUPERSEDED", "LEASE_EXPIRED" ]
          }
        }, {
          "in" : "query",
          "name" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "",
            "maxLength" : 100,
            "minLength" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyAiRunPage"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "조회 조건 오류"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "로그인 필요"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "권한 또는 CSRF 토큰 확인 필요"
          },
          "503" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "조회 실패"
          }
        },
        "security" : [ {
          "memberSession" : [ ]
        } ],
        "summary" : "AI 자동 추출 상태 조회",
        "tags" : [ "policy-ai-run-controller" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "AdminSliceCollectionExceptionItem" : {
        "type" : "object",
        "properties" : {
          "hasNext" : {
            "type" : "boolean"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CollectionExceptionItem"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "1부터 시작하는 페이지"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "hasNext", "items", "page", "pageSize" ]
      },
      "AdminSliceCollectionPageFailure" : {
        "type" : "object",
        "properties" : {
          "hasNext" : {
            "type" : "boolean"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CollectionPageFailure"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "1부터 시작하는 페이지"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "hasNext", "items", "page", "pageSize" ]
      },
      "AdminSliceCollectionReplayResult" : {
        "type" : "object",
        "properties" : {
          "hasNext" : {
            "type" : "boolean"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CollectionReplayResult"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "1부터 시작하는 페이지"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "hasNext", "items", "page", "pageSize" ]
      },
...
      "CollectionExceptionCurrentPolicy" : {
        "type" : "object",
        "properties" : {
          "collectedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "content" : {
            "$ref" : "#/components/schemas/PolicyContent"
          },
          "correctionId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          },
          "policyNumber" : {
            "type" : "string"
          },
          "previousRevision" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/CollectionExceptionRevision"
            }, {
              "type" : "null"
            } ],
            "description" : "같은 정책의 직전 내부 개정. 없으면 null"
          },
          "revision" : {
            "type" : "integer",
            "format" : "int64"
          },
          "sourceCapturedAt" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "현재 개정이 참조하는 원본 수집 시각. 개정 적용 시각이 아님"
          }
        },
        "required" : [ "collectedAt", "content", "correctionId", "policyNumber", "previousRevision", "revision", "sourceCapturedAt" ]
      },
      "CollectionExceptionDetail" : {
        "type" : "object",
        "properties" : {
          "currentPolicy" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/CollectionExceptionCurrentPolicy"
            }, {
              "type" : "null"
            } ],
            "description" : "같은 정책번호의 조회 시점 공개 내용. 없으면 null"
          },
          "item" : {
            "$ref" : "#/components/schemas/CollectionExceptionItem"
          },
          "rawPolicyJson" : {
            "type" : "string",
            "description" : "저장된 수집 항목의 JSON 문자열. 외부 원문이므로 실행하지 않고 텍스트로 표시"
          }
        },
        "required" : [ "currentPolicy", "item", "rawPolicyJson" ]
      },
      "CollectionExceptionItem" : {
        "type" : "object",
        "properties" : {
          "attempts" : {
            "type" : "integer",
            "format" : "int32"
          },
          "itemIndex" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "수집 페이지 내 0부터 시작하는 항목 위치"
          },
          "lastAttemptAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "outcome" : {
            "type" : "string",
            "enum" : [ "INVALID_ITEM", "STORE_FAILED", "CORRECTION_CONFLICT" ]
          },
          "pageNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "policyNumber" : {
            "type" : [ "string", "null" ],
            "description" : "형식을 확인한 원천 정책번호. 확인 불가 시 null"
          },
          "runId" : {
            "type" : "string",
            "format" : "uuid"
          }
        },
        "required" : [ "attempts", "itemIndex", "lastAttemptAt", "outcome", "pageNumber", "policyNumber", "runId" ]
      },
      "CollectionExceptionRevision" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "$ref" : "#/components/schemas/PolicyContent"
          },
          "correctionId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          },
          "revision" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "서비스 내부 개정 번호"
          },
          "sourceCapturedAt" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "개정이 참조하는 원본 수집 시각. 개정 적용 시각이 아님"
          }
        },
        "required" : [ "content", "correctionId", "revision", "sourceCapturedAt" ]
      },
      "CollectionPageFailure" : {
        "type" : "object",
        "properties" : {
          "dispatchedAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "httpStatus" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          },
          "pageNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "reason" : {
            "type" : "string",
            "enum" : [ "API_KEY_MISSING", "HTTP_ERROR", "NON_JSON_RESPONSE", "SECRET_IN_RESPONSE", "REQUEST_INTERRUPTED", "REQUEST_OR_RESPONSE_FAILED", "RESPONSE_STORE_FAILED", "INVALID_LIST_RESPONSE", "UNKNOWN" ]
          },
          "receivedAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "responseStored" : {
            "type" : "boolean"
          },
          "runId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "startedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "state" : {
            "type" : "string",
            "enum" : [ "FETCH_FAILED", "INVALID_RESPONSE" ]
          }
        },
        "required" : [ "dispatchedAt", "httpStatus", "pageNumber", "reason", "receivedAt", "responseStored", "runId", "startedAt", "state" ]
      },
      "CollectionReplayRequest" : {
        "type" : "object",
        "properties" : {
          "expectedAttempts" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          },
          "reason" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          },
          "requestId" : {
            "type" : "string",
            "format" : "uuid"
          }
        },
        "required" : [ "expectedAttempts", "reason", "requestId" ]
      },
      "CollectionReplayResult" : {
        "type" : "object",
        "properties" : {
          "actorId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "attempt" : {
            "type" : "integer",
            "format" : "int32"
          },
          "expectedAttempts" : {
            "type" : "integer",
            "format" : "int32"
          },
          "itemIndex" : {
            "type" : "integer",
            "format" : "int32"
          },
          "outcome" : {
            "type" : "string",
            "enum" : [ "APPLIED", "UNCHANGED", "REPLAYED", "STALE", "INVALID_ITEM", "CORRECTION_CONFLICT" ]
          },
          "policyNumber" : {
            "type" : [ "string", "null" ]
          },
          "policyRevision" : {
            "type" : [ "integer", "null" ],
            "format" : "int64",
            "description" : "처리 후 개정. 반영하지 않은 결과는 null"
          },
          "processedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "reason" : {
            "type" : "string"
          },
          "requestId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "runId" : {
            "type" : "string",
            "format" : "uuid"
          }
        },
        "required" : [ "actorId", "attempt", "expectedAttempts", "itemIndex", "outcome", "policyNumber", "policyRevision", "processedAt", "reason", "requestId", "runId" ]
      },
...
      "PolicyAiRunItem" : {
        "type" : "object",
        "properties" : {
          "attempt" : {
            "type" : "integer",
            "format" : "int32"
          },
          "candidateStatus" : {
            "type" : [ "string", "null" ],
            "description" : "조회 시점의 추출 결과. 마지막 시도 종료 후 저장됐을 수 있다"
          },
          "currentRevision" : {
            "type" : "integer",
            "format" : "int64"
          },
          "finishedAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "latestRequest" : {
            "type" : "boolean"
          },
          "policyNumber" : {
            "type" : "string"
          },
          "requestId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "reservationPhase" : {
            "type" : [ "string", "null" ],
            "description" : "조회 시점의 비용 예약 상태"
          },
          "responseStored" : {
            "type" : "boolean"
          },
          "resultCode" : {
            "type" : [ "string", "null" ],
            "description" : "마지막 시도 종료 시 기록한 처리 코드"
          },
          "revision" : {
            "type" : "integer",
            "format" : "int64"
          },
          "sourceMatches" : {
            "type" : "boolean"
          },
          "startedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "state" : {
            "type" : "string",
            "description" : "요청별 마지막 자동 시도 상태. 실행 기한이 지난 RUNNING은 LEASE_EXPIRED로 표시하며 DB는 변경하지 않는다",
            "enum" : [ "RUNNING", "COMPLETED", "RETRY_PENDING", "INTERRUPTED", "REVIEW_REQUIRED", "SUPERSEDED", "LEASE_EXPIRED" ]
          },
          "title" : {
            "type" : "string"
          }
        },
        "required" : [ "attempt", "candidateStatus", "currentRevision", "finishedAt", "latestRequest", "policyNumber", "requestId", "reservationPhase", "responseStored", "resultCode", "revision", "sourceMatches", "startedAt", "state", "title" ]
      },
      "PolicyAiRunPage" : {
        "type" : "object",
        "properties" : {
          "automationEnabled" : {
            "type" : "boolean",
            "description" : "자동 실행 스위치. AI 설정·예산·일일 한도 충족 여부를 뜻하지 않는다"
          },
          "checkedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "hasNext" : {
            "type" : "boolean"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PolicyAiRunItem"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "automationEnabled", "checkedAt", "hasNext", "items", "page", "pageSize", "total" ]
      },
...
      "PolicyApiError" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "message" ]
      },
...
      "PolicyContent" : {
        "type" : "object",
        "properties" : {
          "applicationPeriod" : {
            "type" : "string"
          },
          "category" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "links" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PolicyOfficialLink"
            }
          },
          "organization" : {
            "type" : "string"
          },
          "regionCodes" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "sections" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PolicyTextSection"
            }
          },
          "sourceModifiedAtText" : {
            "type" : "string",
            "description" : "시간대가 확정되지 않은 원천 수정 일시 원문"
          },
          "title" : {
            "type" : "string"
          }
        },
        "required" : [ "applicationPeriod", "category", "description", "links", "organization", "regionCodes", "sections", "sourceModifiedAtText", "title" ]
      },
...
    },
    "securitySchemes" : {
      "memberSession" : {
        "description" : "카카오·네이버 로그인 후 발급하는 서버 세션",
        "in" : "cookie",
        "name" : "YPM_SESSION",
        "type" : "apiKey"
      }
    }
  }
}